Add keycodes for virtual keys on 7x30 fluid device
Fluid board uses the keypad layout file "fluid-keypad.kl" since
the corresponding keypad input device is named as fluid-keypad
in the kernel. Android processes keys only when they are
part of this layout file. Though, virtual keys are
part of touchscreen they will not be processed unless these
keycodes are part of fluid-keypad.kl. So, add keycodes for BACK,
SEARCH and MENU which are part of touchscreen's virtual keys

Change-Id: Ie1e3547a471887f27f1218c9969cc9e9a5a01eef